backend/requirements-gpu.txt
Flask==2.3.3
diffusers==0.20.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
transformers==4.30.2
accelerate==0.20.3
requests==2.31.0
Pillow==10.0.0
